High-level API reference

This is the official High-level API reference of JustSayIt. Note that it can also be queried interactively from the Julia REPL using the help mode:

julia> using JustSayIt
julia>?
help?> JustSayIt.API

Submodule

JustSayIt.API — Module

Module API

Application Programming Interface (API) of JustSayIt.

Macros

@voiceargs

@voiceconfig

@voiceinfo

Functions

Lifecycle and device access: controller, init_jsi, finalize_jsi, default_language, type_languages

Interpreters and STT helpers: interpret_enum, interpret_digit, interpret_count, interpret_language, next_wordgroup, next_letter, next_letters, next_digit, next_digits, get_language

Text access: get_selection_content, get_clipboard_content

TTS helpers: is_playing_tts, pause_tts, resume_tts, stop_tts, set_tts_async_default, tts_async_default

LLM helpers: ask_llm, ask_llm!

Constants

LANG

LANG_AUTO

MODELTYPE_DEFAULT

MODELTYPE_SPEECH

ALPHABET

DIGITS

COUNTS

To see a description of a function, macro or module type ?<functionname>, ?<macroname> (including the @) or ?<modulename>, respectively.

Macros

Index

JustSayIt.API.@voiceargs

JustSayIt.API.@voiceconfig

JustSayIt.API.@voiceinfo

Documentation

JustSayIt.@voiceargs — Macro

@voiceargs args function

Declare some or all arguments of the function definition to be arguments that can be obtained by voice.

Arguments

args: a voicearg, a pair voicearg=>kwargs or a tuple with multiple voiceargs, which can each have kwargs or not (see examples below).

function: the function definition.

Keyword arguments definable for each voice argument in `args`

model::String=MODELNAME.DEFAULT.<default_language>: the name of the model to be used for the function argument (the name must be one of the keys of the modeldirs dictionary passed to init_jsi).

modeltype::String=MODELNAME.DEFAULT: the name of the model type to be used for the function argument (a IETF language tag, elsewhere selected, will be concatenated to the model type to form the model name; the model names constructed in this manner must be keys of the modeldirs dictionary passed to init_jsi).

language::String=<default_language>: the language (IETF language tag) to be used for the function argument (the model name will be constructed by concatenating the model type and the language).

valid_input::AbstractVector{String}|NTuple{N,Pair{String,<:AbstractVector{String}}}|Dict{String,<:AbstractVector{String}: the valid speech input (e.g. ["up", "down"] or ("de" => ["rauf", "runter"], "en-us" => ["up", "down"], "fr" => ["haut", "bas"], "es" => ["arriba", "abajo"])).

valid_input_auto::Bool: whether the valid speech input can automatically be derived from the type of the function argument.

interpreter::Function: a function to interpret the token (mapping a String to a different String).

use_max_speed::Bool=false: whether to use maxium speed for the recognition of the next token (rather than maximum accuracy). It is generally only recommended to set use_max_speed=true for single word commands or very specfic use cases that require immediate minimal latency action when a command is said.

ignore_unknown::Bool=false: whether to ignore unknown tokens in the speech (and consume the next instead). It is generally not recommended to set ignore_unkown=true - in particular not in combination with a limited valid input - as then the function will block until it receives a token it recognizes.

timeout::AbstractFloat: timeout (in seconds) after which to abort argument recognition.

vararg_end::String: a token to signal the end of a vararg (only valid if the function argument is a vararg).

vararg_max::Integer=∞: the maximum number of arguments the vararg can contain (only valid if the function argument is a vararg).

Examples

@voiceargs (b, c) function f(a, b::String, c::String, d)
    #(...)
    return
end

@voiceargs (b=>(use_max_speed=true), c) function f(a, b::String, c::String, d)
    #(...)
    return
end

@enum TypeMode words formula
@voiceargs (mode=>(valid_input_auto=true), token=>(model=MODELNAME.TYPE.EN_US, timeout=2.0)) function type_tokens(mode::TypeMode, tokens::String...)
    #(...)
    return
end

See also: @voiceconfig

JustSayIt.@voiceconfig — Macro

@voiceconfig kwargs call
@voiceconfig kwargs symbol

Define voice recognition keyword arguments for all arguments of a function call or function symbol; recognition keyword arguments defined explicitly in a @voiceargs function definition precede those defined with @voiceconfig. See the @voiceargs macro for the valid keyword arguments.

Arguments

kwargs: a keyword argument or a named tuple of keyword arguments (e.g., language="de" or (language="de", timeout=10.0)).

call: a function call expression (e.g., f(x, y=3)) or a function symbol (e.g., f).

symbol: a function symbol (e.g., f).

Examples

@voiceconfig language="de" f(x; y=3)

@voiceconfig language="de" f

@voiceconfig (language="de", timeout=10.0) f(x; y=3)

@voiceconfig (language="de", timeout=10.0) f

See also: @voiceargs

JustSayIt.TTScore.@voiceinfo — Macro

@voiceinfo message

Log message and speak it through the active TTS engine.

Functions

Index

JustSayIt.API.finalize_jsi

JustSayIt.API.init_jsi

JustSayIt.API.controller

JustSayIt.API.interpret_enum

JustSayIt.API.interpret_digit

JustSayIt.API.interpret_count

JustSayIt.API.interpret_language

JustSayIt.API.default_language

JustSayIt.API.type_languages

JustSayIt.API.ask_llm

JustSayIt.API.ask_llm!

JustSayIt.API.next_wordgroup

JustSayIt.API.next_letter

JustSayIt.API.next_letters

JustSayIt.API.next_digit

JustSayIt.API.next_digits

JustSayIt.API.get_language

JustSayIt.API.get_selection_content

JustSayIt.API.get_clipboard_content

JustSayIt.API.is_playing_tts

JustSayIt.API.pause_tts

JustSayIt.API.resume_tts

JustSayIt.API.stop_tts

JustSayIt.API.set_tts_async_default

JustSayIt.API.tts_async_default

Documentation

JustSayIt.controller — Function

controller(name)

Return the configured device controller named name, for example keyboard or mouse.

JustSayIt.finalize_jsi — Function

finalize_jsi()

Finalize the package JustSayIt.

See also: init_jsi

JustSayIt.init_jsi — Function

init_jsi()
init_jsi(<keyword arguments>)

Initialize the package JustSayIt.

Keyword arguments

default_language::String="en-us": the default language, which is used for the voice arguments and for typing when no other language is specified (noted with its IETF language tag https://en.wikipedia.org/wiki/IETFlanguagetag). Currently supported are: english-US ("en-us"), German ("de"), French ("fr") and Spanish ("es").

type_languages::String|AbstractArray{String}=default_language: the languages used for typing, where the first is the default type language (noted with its IETF language tag https://en.wikipedia.org/wiki/IETFlanguagetag). Currently supported are: english-US ("en-us"), German ("de"), French ("fr") and Spanish ("es"). Type ?Keyboard.type for information about typing or say "help type" after having started JustSayIt.

use_llm::Bool=true: whether to use a Large Language Model (LLM) for advanced features as the summary or translation of text (default is `true).

use_tts::Bool=true: whether to use text-to-speech (TTS) functionality for voice enhancement of certain commands and for enabling the text-to-speech commands of the model TTS (default is true).

tts_async_default::Bool=true: whether to use asynchronous TTS by default. The default (true) requires using of headphones, because in asynchronous mode, the TTS voice can naturally trigger commands by accident.

use_gpu::Bool=true: whether to use GPU, enabling the use of better performing STT, TTS and LLM models (default is true).

microphone_id::Int: the id of the microphone to be used, instead of the default microphone. JustSayIt prints available devices and their ids at the start of the application.

microphone_name::String: the name of the microphone to be used, instead of the default microphone. JustSayIt prints available devices at the start of the application (names are sometimes better used instead of ids, because the latter can change dynamically in certain OS). The name is case-insensitive and must match the beginning of the printed device name (if there are multiple matches, the first is used).

audiooutput_id::Int: the id of the audio output device to be used, instead of the default audio output device. JustSayIt prints available devices and their ids at the start of the application.

audiooutput_name::String: the name of the audio output device to be used, instead of the default audio output device. JustSayIt prints available devices at the start of the application (names are sometimes better used instead of ids, because the latter can change dynamically in certain OS). The name is case-insensitive and must match the beginning of the printed device name (if there are multiple matches, the first is used).

Advanced

modeldirs::Dict{String, String}: the directories where the unziped speech recognition models to be used are located. If modeldirs is not set, then it is automatically defined according to the default_language and type_languages set. Models are downloadable from here: https://alphacephei.com/vosk/models

noises::Dict{String, <:AbstractArray{String}}=DEFAULT_NOISES: for each model, an array of strings with noises (tokens that are to be ignored in the speech as, e.g., "huh").

audio_input_cmd::Cmd=nothing: a command that returns an audio stream to replace the default audio recorder. The audio stream must fullfill the following properties: samplerate=16000, channels=1 and format=Int16 (signed 16-bit integer).

Examples

Define custom modeldirs

using JustSayIt
modeldirs = Dict(MODELNAME.DEFAULT.EN_US => "/home/runner/mymodels/vosk-model-small-en-us-0.15",
                 MODELNAME.TYPE.EN_US    => "/home/runner/mymodels/vosk-model-en-us-daanzu-20200905",
                 MODELNAME.DEFAULT.FR    => "/home/runner/mymodels/vosk-model-small-fr-0.22",
                 MODELNAME.TYPE.FR       => "/home/runner/mymodels/vosk-model-fr-0.6-linto-2.2.0")
start(modeldirs=modeldirs, default_language="en-us", type_languages=["en-us","fr"])

Define audio_input_cmd

# Use a custom command to create the audio input stream - instead of the default recorder (the rate, channels and format must not be chosen different!)
using JustSayIt
audio_input_cmd = `arecord --rate=16000 --channels=1 --format=S16_LE`
start(audio_input_cmd=audio_input_cmd)

Default Vosk model directories

__default__-de    => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-small-de-0.15
__default__-en-us => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-small-en-us-0.15
__default__-es    => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-small-es-0.22
__default__-fr    => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-small-fr-0.22
__speech__-de     => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-de-0.21
__speech__-en-us  => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-en-us-daanzu-20200905
__speech__-es     => 
__speech__-fr     => /home/runner/.local/share/JustSayIt/models/vosk/vosk-model-fr-0.22

Default noises

__default__-de    => String[]
__default__-en-us => ["huh"]
__default__-es    => String[]
__default__-fr    => ["hum"]
__speech__-auto   => String[]
__speech__-de     => String[]
__speech__-en-us  => ["huh"]
__speech__-es     => String[]
__speech__-fr     => ["hum"]

JustSayIt.interpret_enum — Function

interpret_enum(input, valid_input)

Map the localized spoken input found in valid_input to the corresponding English enum label.

JustSayIt.interpret_digit — Function

Get the digit symbol corresponding to the spoken input in the current default language.

JustSayIt.interpret_count — Function

Get the count symbol corresponding to the spoken input in the current default language.

JustSayIt.interpret_language — Function

Get the IETF language tag corresponding to the spoken input in the current default language.

JustSayIt.default_language — Function

default_language()

Return the current default language used for speech recognition and typing.

JustSayIt.type_languages — Function

type_languages()

Return the currently configured typing languages, ordered by priority.

JustSayIt.LLMcore.ask_llm — Function

ask_llm(question)

Query the configured LLM with question and return the answer text.

See also: ask_llm!

JustSayIt.LLMcore.ask_llm! — Function

ask_llm!(question)

Query the configured LLM with question as a follow-up and return the answer text.

See also: ask_llm

JustSayIt.next_wordgroup — Function

Get next word group from speech.

JustSayIt.next_letter — Function

Get next letter from speech.

JustSayIt.next_letters — Function

Get next letters from speech.

JustSayIt.next_digit — Function

Get next digit from speech.

JustSayIt.next_digits — Function

Get next digits from speech.

JustSayIt.get_language — Function

Get the language (IETF language tag) from speech.

JustSayIt.get_selection_content — Function

get_selection_content()

Get the current primary text selection.

JustSayIt.get_clipboard_content — Function

get_clipboard_content()

Get the current clipboard text content.

JustSayIt.TTScore.is_playing_tts — Function

Return whether TTS playback is currently active on the selected engine and stream.

JustSayIt.TTScore.pause_tts — Function

Pause the current TTS playback on the selected engine and stream.

JustSayIt.TTScore.resume_tts — Function

Resume paused TTS playback on the selected engine and stream.

JustSayIt.TTScore.stop_tts — Function

Stop the current TTS playback on the selected engine and stream.

JustSayIt.TTScore.set_tts_async_default — Function

set_tts_async_default(async)

Set whether JustSayIt uses asynchronous TTS playback by default.

JustSayIt.TTScore.tts_async_default — Function

tts_async_default()

Return whether JustSayIt uses asynchronous TTS playback by default.

Constants

Index

JustSayIt.API.LANG

JustSayIt.API.LANG_AUTO

JustSayIt.API.MODELTYPE_DEFAULT

JustSayIt.API.MODELTYPE_SPEECH

JustSayIt.API.ALPHABET

JustSayIt.API.DIGITS

JustSayIt.API.COUNTS

Documentation

JustSayIt.LANG — Constant

Constant named tuple containing names of available languages.

JustSayIt.LANG_AUTO — Constant

Constant language tag used for automatic language detection.

JustSayIt.MODELTYPE_DEFAULT — Constant

Constant model-type prefix for command-oriented speech-recognition models.

JustSayIt.MODELTYPE_SPEECH — Constant

Constant model-type prefix for free-speech recognition models.

JustSayIt.ALPHABET — Constant

Constant mapping from each supported language to the recognized spoken letter tokens.

JustSayIt.DIGITS — Constant

Constant mapping from each supported language to the recognized spoken digit tokens.

JustSayIt.COUNTS — Constant

Constant mapping from each supported language to the recognized spoken count tokens.