Software reference

This is the official Software reference of JustSayIt. Note that it can also be queried interactively from the Julia REPL using the help mode:

julia> using JustSayIt
julia>?
help?> JustSayIt

Root module

JustSayIt.JustSayIt — Module

Module JustSayIt

Enables offline, low latency, highly accurate speech to command translation, offering composable functionality using speech-to-text (STT), text-to-speech (TTS), large language models (LLMs) and more. It contains in addition an Application Programming Interface (API) for a fast and easy development of custom functions.

General overview and examples

https://github.com/omlins/JustSayIt.jl

Software

> julia
julia> using JustSayIt
julia> commands = Dict(
    ...
)
julia> start(commands=commands, <keyword arguments>)

Type ?start to learn about customization keywords to pass a command dictionary and set different options.

API

> julia
julia> import JustSayIt
julia> using JustSayIt.API

Type ?JustSayIt.API to learn about the Application Programming Interface (JustSayIt.API) of JustSayIt.

Functions to configure and start JustSayIt

Index

JustSayIt.start

JustSayIt.@include

JustSayIt.@edit

Documentation

JustSayIt.start — Function

start(<keyword arguments>)

Start offline, low latency, highly accurate and secure speech to command translation.

Keyword arguments

commands::Dict{String, <:Any}=DEFAULT_COMMANDS[default_language]: the commands to be recognized with their mapping to a function or to a keyboard key or shortcut or a sequence of any of those. If you define your own top-level command dictionary and want spoken help there as well, include the language-appropriate help command explicitly (for example "help" => Help.help for english-US).

subset::AbstractArray{String}=nothing: a subset of the commands to be recognised and executed (instead of the complete commands list).

max_speed_subset::AbstractArray{String}=nothing: a subset of the commands for which the command names (first word of a command) are to be recognised with maxium speed rather than with maximum accuracy. Forcing maximum speed is usually desired for single word commands that map to functions or keyboard shortcuts that should trigger immediate actions as, e.g., mouse clicks or page up/down (in general, actions that do not modify content and can therefore safely be triggered at maximum speed). Note that forcing maximum speed means not to wait for a certain amount of silence after the end of a command as normally done for the full confirmation of a recognition. As a result, it enables a minimal latency between the saying of a command name and its execution. Note that it is usually possible to define very distinctive command names, which allow for a safe command name to shortcut mapping at maximum speed (to be tested case by case).

default_language::String="en-us": the default language, which is used for the command names, for the voice arguments and for typing when no other language is specified (noted with its IETF language tag https://en.wikipedia.org/wiki/IETFlanguagetag). Currently supported are: english-US ("en-us"), German ("de"), French ("fr") and Spanish ("es").

type_languages::String|AbstractArray{String}=default_language: the languages used for typing, where the first is the default type language (noted with its IETF language tag https://en.wikipedia.org/wiki/IETFlanguagetag). Currently supported are: english-US ("en-us"), German ("de"), French ("fr") and Spanish ("es"). Type ?Keyboard.type for information about typing or say "help type" after having started JustSayIt.

use_llm::Bool=true: whether to use a Large Language Model (LLM) for advanced features as the summary or translation of text (default is `true).

use_tts::Bool=true: whether to use text-to-speech (TTS) functionality for voice enhancement of certain commands and for enabling the text-to-speech commands of the model TTS (default is true).

tts_async_default::Bool=true: whether to use asynchronous TTS by default. The default (true) requires using of headphones, because in asynchronous mode, the TTS voice can naturally trigger commands by accident.

use_gpu::Bool=true: whether to use GPU, enabling the use of better performing STT, TTS and LLM models (default is true).

microphone_id::Int: the id of the microphone to be used, instead of the default microphone. JustSayIt prints available devices and their ids at the start of the application.

microphone_name::String: the name of the microphone to be used, instead of the default microphone. JustSayIt prints available devices at the start of the application (names are sometimes better used instead of ids, because the latter can change dynamically in certain OS). The name is case-insensitive and must match the beginning of the printed device name (if there are multiple matches, the first is used).

audiooutput_id::Int: the id of the audio output device to be used, instead of the default audio output device. JustSayIt prints available devices and their ids at the start of the application.

audiooutput_name::String: the name of the audio output device to be used, instead of the default audio output device. JustSayIt prints available devices at the start of the application (names are sometimes better used instead of ids, because the latter can change dynamically in certain OS). The name is case-insensitive and must match the beginning of the printed device name (if there are multiple matches, the first is used).

Advanced

audio_input_cmd::Cmd=nothing: a command that returns an audio stream to replace the default audio recorder. The audio stream must fullfill the following properties: samplerate=16000, channels=1 and format=Int16 (signed 16-bit integer).

Submodules for command name to function mapping (alphabetically ordered)

Clipboard

Generic

Help

Keyboard

LLM

Mouse

Selection

STT

TTS

To see a description of a submodule, type ?<modulename>.

Default commands

de    => Dict("hilfe" => JustSayIt.Help.help)
en-us => Dict("help" => JustSayIt.Help.help)
es    => Dict("ayuda" => JustSayIt.Help.help)
fr    => Dict("aide" => JustSayIt.Help.help)

Examples

Define commands - functions, keyboard shortcuts and sequences of those - and a max_speed_subset

using JustSayIt
commands = Dict("help"      => Help.help,
                "ma"        => Mouse.click_left,
                "select"    => Mouse.press_left,
                "okay"      => Mouse.release_left,
                "middle"    => Mouse.click_middle,
                "right"     => Mouse.click_right,
                "double"    => Mouse.click_double,
                "triple"    => Mouse.click_triple,
                "copy"      => (Key.ctrl, 'c'),
                "cut"       => (Key.ctrl, 'x'),
                "paste"     => (Key.ctrl, 'v'),
                "undo"      => (Key.ctrl, 'z'),
                "redo"      => (Key.ctrl, Key.shift, 'z'),
                "upwards"   => Key.page_up,
                "downwards" => Key.page_down,
                "take"      => [Mouse.click_double, (Key.ctrl, 'c')],
                "replace"   => [Mouse.click_double, (Key.ctrl, 'v')],
                );
start(commands=commands, max_speed_subset=["ma", "select", "okay", "middle", "right", "double", "triple", "copy", "upwards", "downwards", "take"])

Define commands with nested subcommands - activated automatically when the parent command is said

using JustSayIt
vsc_commands = Dict("comment" => (Key.ctrl, '/'),
                    "save"    => (Key.ctrl, 's'))
commands = Dict("help"   => Help.help,
                "coding" => [`code`, vsc_commands])
start(commands=commands)

Define default_language (also used for typing if type_languages is not set)

# Set command and type language to french:
using JustSayIt
start(default_language="fr")

Define default_language and type_languages

# Set command language to french and type languages to french and spanish:
using JustSayIt
start(default_language="fr", type_languages=["fr","es"])

JustSayIt.@include — Macro

@include(path::AbstractString)

Prefix path with the JustSayIt application config path and then call include(path). If path is an absolut path, then call include with path unmodified.

NOTE: JustSayIt application config

The content of the JustSayIt application config folder is not evaluated within JustSayIt. The folder's single purpose is to provide an easily accessible storage for scripts to start JustSayIt and/or for custom command functions: @include permits to conveniently include files from this folder (for details about the Julia built-in include type ?include). Your JustSayIt application config path on this system is: /home/runner/.config/JustSayIt

JustSayIt.@edit — Macro

@edit(path::AbstractString)

Prefix path with the JustSayIt application config path and then call edit(path). If path is an absolut path, then call edit with path unmodified.

NOTE: JustSayIt application config

The content of the JustSayIt application config folder is not evaluated within JustSayIt. The folder's single purpose is to provide an easily accessible storage for scripts to start JustSayIt and/or for custom command functions: @edit permits to conveniently edit files from this folder (for details about the Julia built-in edit type ?edit). Your JustSayIt application config path on this system is: /home/runner/.config/JustSayIt

Keyboard shortcut namespace

The root JustSayIt module also re-exports Key for command mappings and keyboard shortcuts. Use Key. completion in the Julia REPL to inspect the currently available key names.

Composable primitives for quick command definition (reexported from the root module for convenient access)

These root primitives can be composed directly in command dictionaries when no wrapper function is needed.

Use type here for low-level text typing when you already have text to send to the active application. For voice-input typing from speech, use the Keyboard.type_* family documented below.

Index

JustSayIt.take

JustSayIt.type

JustSayIt.ask

JustSayIt.ask!

JustSayIt.grab

JustSayIt.listen

JustSayIt.say

Documentation

JustSayIt.Clipboard.take — Function

take

Take the content of the clipboard.

JustSayIt.Keyboard.type — Function

type(text::AbstractString)

Type text with the keyboard.

JustSayIt.LLM.ask — Function

ask(question::AbstractString)

Use LLM to reply to the question.

JustSayIt.LLM.ask! — Function

ask!(question::AbstractString)

Use LLM to reply to the question and type the answer.

JustSayIt.Selection.grab — Function

grab

Grab the text selected by the user.

JustSayIt.STT.listen — Function

Listen to speech and return the next word group.

JustSayIt.TTScore.say — Function

say(text)

Read text with the active TTS engine.

Submodules for quick command definition with predefined, flexible and composable functions

Index

JustSayIt.Clipboard

JustSayIt.Generic

JustSayIt.Help

JustSayIt.Keyboard

JustSayIt.LLM

JustSayIt.Mouse

JustSayIt.Selection

JustSayIt.STT

JustSayIt.TTS

Documentation

JustSayIt.Clipboard — Module

Module Clipboard

Provides functions for accessing the clipboard.

Functions

Clipboard access

Clipboard.take

To see a description of a function type ?<functionname>.

See also: JustSayIt.Selection

JustSayIt.Generic — Module

Module Generic

Provides generic functions for any kind of operations.

Functions

Command execution

Generic.execute_n_times

To see a description of a function type ?<functionname>.

JustSayIt.Help — Module

Module Help

Provides functions for displaying help about the commands during speech recognition.

Functions

Help.help

To see a description of a function type ?<functionname>.

JustSayIt.Keyboard — Module

Module Keyboard

Provides functions for controlling the keyboard by voice.

Functions

Text typing

Keyboard.type_sentence

Keyboard.type_sentence_lowercase

Word typing

Keyboard.type_first_lowercase

Keyboard.type_first_uppercase

Keyboard.type_lowercase

Keyboard.type_uppercase

Keyboard.type_flatcase

Keyboard.type_snakecase

Keyboard.type_camelcase

Keyboard.type_constantcase

Letter and digit typing

Keyboard.type_letter

Keyboard.type_majuscule

Keyboard.type_letters

Keyboard.type_capitals

Keyboard.type_digits

Key pressing

Keyboard.press_keys

Keyboard.press_delete

Keyboard.press_backspace

String typing

Keyboard.type

To see a description of a function type ?<functionname>.

See also: JustSayIt.Mouse

JustSayIt.LLM — Module

Module LLM

Provides functions for operations using an LLM.

Functions

Text writing

LLM.type_summary

LLM.type_translation

LLM.type_answer

LLM.type_text_answer

Text reading

LLM.read_summary

LLM.read_translation

LLM.read_answer

LLM.read_text_answer

Text generation (no typing or reading)

LLM.ask

LLM.ask!

To see a description of a function type ?<functionname>.

JustSayIt.Mouse — Module

Module Mouse

Provides functions for controlling the mouse by voice.

Functions

Button control

Mouse.click_left

Mouse.click_middle

Mouse.click_right

Mouse.press_left

Mouse.release_left

Mouse.click_double

Mouse.click_triple

To see a description of a function type ?<functionname>.

See also: JustSayIt.Keyboard

JustSayIt.Selection — Module

Module Selection

Provides functions for accessing the text selected by the user.

Functions

Selection Access

Selection.grab

To see a description of a function type ?<functionname>.

See also: JustSayIt.Clipboard

JustSayIt.STT — Module

Module STT

Provides functions for speech-to-text (STT).

Functions

Text recognition

STT.listen

To see a description of a function type ?<functionname>.

JustSayIt.TTS — Module

Module TTS

Provides functions for text-to-speech (TTS).

Functions

Text reading

TTS.read

TTS.pause

TTS.resume

TTS.stop

TTS.set_async_default

String reading

TTS.say

To see a description of a function type ?<functionname>.

Functions of Clipboard

Index

JustSayIt.Clipboard.take

JustSayIt.Clipboard.take

Documentation

JustSayIt.Clipboard.take — Method

take

Take the content of the clipboard.

Functions of Generic

Index

JustSayIt.Generic.execute_n_times

Documentation

JustSayIt.Generic.execute_n_times — Function

execute_n_times `n` `(complement of) cmd`

Execute a command n times.

Functions of Help

Index

JustSayIt.Help.help

Documentation

JustSayIt.Help.help — Method

Show help for your commands or a spoken command or module.

Functions of Keyboard

Index

JustSayIt.Keyboard.press_backspace

JustSayIt.Keyboard.press_delete

JustSayIt.Keyboard.press_keys

JustSayIt.Keyboard.type

JustSayIt.Keyboard.type

JustSayIt.Keyboard.type_camelcase

JustSayIt.Keyboard.type_capitals

JustSayIt.Keyboard.type_constantcase

JustSayIt.Keyboard.type_digits

JustSayIt.Keyboard.type_first_lowercase

JustSayIt.Keyboard.type_first_uppercase

JustSayIt.Keyboard.type_flatcase

JustSayIt.Keyboard.type_letter

JustSayIt.Keyboard.type_letters

JustSayIt.Keyboard.type_lowercase

JustSayIt.Keyboard.type_majuscule

JustSayIt.Keyboard.type_sentence

JustSayIt.Keyboard.type_sentence_lowercase

JustSayIt.Keyboard.type_snakecase

JustSayIt.Keyboard.type_uppercase

Documentation

JustSayIt.Keyboard.press_backspace — Method

Press backspace key.

JustSayIt.Keyboard.press_delete — Method

Press delete key.

JustSayIt.Keyboard.press_keys — Method

press_keys(keys...; count=1)

Press keys on the keyboard.

JustSayIt.Keyboard.type — Method

type(text::AbstractString)

Type text with the keyboard.

JustSayIt.Keyboard.type_camelcase — Method

Type words in camelcase (each word's first letter uppercase and squashed together without separator).

JustSayIt.Keyboard.type_capitals — Method

Type capitals (exit on unknown).

JustSayIt.Keyboard.type_constantcase — Method

Type words in constantcase (uppercase, separated with underscore).

JustSayIt.Keyboard.type_digits — Method

Type digits (exit on unknown).

JustSayIt.Keyboard.type_first_lowercase — Method

Type words, starting with the first in lowercase (separated with space).

JustSayIt.Keyboard.type_first_uppercase — Method

Type first words, starting with the first in uppercase (separated with space).

JustSayIt.Keyboard.type_flatcase — Method

Type words in flatcase (lowercase, squashed together without separator).

JustSayIt.Keyboard.type_letter — Method

Type one letter.

JustSayIt.Keyboard.type_letters — Method

Type letters (exit on unknown).

JustSayIt.Keyboard.type_lowercase — Method

Type words in all lowercase (separated with space).

JustSayIt.Keyboard.type_majuscule — Method

Type one majuscule letter.

JustSayIt.Keyboard.type_sentence — Method

Type a sentence (starting uppercase).

JustSayIt.Keyboard.type_sentence_lowercase — Method

Type a sentence (starting lowercase).

JustSayIt.Keyboard.type_snakecase — Method

Type words in snakecase (lowercase, separated with underscore).

JustSayIt.Keyboard.type_uppercase — Method

Type words in all uppercase (separated with space).

Functions of LLM

Index

JustSayIt.LLM.ask

JustSayIt.LLM.ask

JustSayIt.LLM.ask!

JustSayIt.LLM.ask!

JustSayIt.LLM.chat_followup

JustSayIt.LLM.read_answer

JustSayIt.LLM.read_followup

JustSayIt.LLM.read_summary

JustSayIt.LLM.read_text_answer

JustSayIt.LLM.read_text_answer_to

JustSayIt.LLM.read_translation

JustSayIt.LLM.type_answer

JustSayIt.LLM.type_followup

JustSayIt.LLM.type_summary

JustSayIt.LLM.type_text_answer

JustSayIt.LLM.type_text_answer_to

JustSayIt.LLM.type_translation

Documentation

JustSayIt.LLM.ask! — Method

ask!(question::AbstractString)

Use LLM to reply to the question and type the answer.

JustSayIt.LLM.ask — Method

ask(question::AbstractString)

Use LLM to reply to the question.

JustSayIt.LLM.chat_followup — Method

chat `question/instructions`...

Use LLM to follow up on the last question and start a chat; say "bye" to exit.

JustSayIt.LLM.read_answer — Method

read_answer `question/instructions`

Use LLM to reply to the question/instructions and read it out loud.

JustSayIt.LLM.read_followup — Method

read_followup `question/instructions`

Use LLM to reply to the follow up question/instructions and read it out loud.

JustSayIt.LLM.read_summary — Method

read_summary

Use LLM to summarize the selected text (or text from clipboard) and read it out loud.

JustSayIt.LLM.read_text_answer — Method

read_text_answer `question/instructions`

Use LLM to reply to the question/instructions considering the selected text (or text from clipboard) and read it out loud.

JustSayIt.LLM.read_text_answer_to — Method

read_text_answer_to(question::AbstractString)

Use LLM to reply to the question/instructions considering the selected text (or text from clipboard) and read it out loud.

JustSayIt.LLM.read_translation — Method

read_translation `language`

Use LLM to translate the selected text (or text from clipboard) to the specified language and read it out loud.

JustSayIt.LLM.type_answer — Method

type_answer `question/instructions`

Use LLM to reply to the question/instructions and type it.

JustSayIt.LLM.type_followup — Method

type_followup `question/instructions`

Use LLM to reply to the follow up question/instructions and type it.

JustSayIt.LLM.type_summary — Method

type_summary

Use LLM to summarize the selected text (or text from clipboard) and type it.

JustSayIt.LLM.type_text_answer — Method

type_text_answer `question/instructions`

Use LLM to reply to the question/instructions considering the selected text (or text from clipboard) and type it.

JustSayIt.LLM.type_text_answer_to — Method

type_text_answer_to(question/instructions)

Use LLM to reply to the question/instructions considering the selected text (or text from clipboard) and type it.

JustSayIt.LLM.type_translation — Method

type_translation `language`

Use LLM to translate the selected text (or text from clipboard) to the specified language and type it.

Functions of Mouse

Index

JustSayIt.Mouse.click_double

JustSayIt.Mouse.click_left

JustSayIt.Mouse.click_middle

JustSayIt.Mouse.click_right

JustSayIt.Mouse.click_triple

JustSayIt.Mouse.move_to_center

JustSayIt.Mouse.move_to_east

JustSayIt.Mouse.move_to_north

JustSayIt.Mouse.move_to_northeast

JustSayIt.Mouse.move_to_northwest

JustSayIt.Mouse.move_to_south

JustSayIt.Mouse.move_to_southeast

JustSayIt.Mouse.move_to_southwest

JustSayIt.Mouse.move_to_west

JustSayIt.Mouse.press_left

JustSayIt.Mouse.release_left

Documentation

JustSayIt.Mouse.click_double — Method

Doubleclick left mouse button.

JustSayIt.Mouse.click_left — Method

Click left mouse button.

JustSayIt.Mouse.click_middle — Method

Click middle mouse button.

JustSayIt.Mouse.click_right — Method

Click right mouse button.

JustSayIt.Mouse.click_triple — Method

Trippleclick left mouse button.

JustSayIt.Mouse.move_to_center — Method

Move to center.

JustSayIt.Mouse.move_to_east — Method

Move to east.

JustSayIt.Mouse.move_to_north — Method

Move to north.

JustSayIt.Mouse.move_to_northeast — Method

Move to north-east.

JustSayIt.Mouse.move_to_northwest — Method

Move to north-west.

JustSayIt.Mouse.move_to_south — Method

Move to south.

JustSayIt.Mouse.move_to_southeast — Method

Move to south-east.

JustSayIt.Mouse.move_to_southwest — Method

Move to south-west.

JustSayIt.Mouse.move_to_west — Method

Move to west.

JustSayIt.Mouse.press_left — Method

Press and hold left mouse button.

JustSayIt.Mouse.release_left — Method

Release left mouse button.

Functions of Selection

Index

JustSayIt.Selection.grab

JustSayIt.Selection.grab

Documentation

JustSayIt.Selection.grab — Method

grab

Grab the text selected by the user.

Functions of STT

Index

JustSayIt.STT.listen

JustSayIt.STT.listen

Documentation

JustSayIt.STT.listen — Method

Listen to speech and return the next word group.

Functions of TTS

Index

JustSayIt.TTS.pause

JustSayIt.TTS.read

JustSayIt.TTS.resume

JustSayIt.TTS.say

JustSayIt.TTS.set_async_default

JustSayIt.TTS.stop

Documentation

JustSayIt.TTS.pause — Method

pause

Pause text reading.

JustSayIt.TTS.read — Method

read

Use TTS to read the selected text (or text from clipboard if no text is selected).

JustSayIt.TTS.resume — Method

resume

Resume text reading.

JustSayIt.TTS.say — Method

say(text::AbstractString)

Use TTS to say the given text.

JustSayIt.TTS.set_async_default — Function

set_async_default `async`

Set the default behavior for TTS to be asynchronous or not.

JustSayIt.TTS.stop — Method

stop

Stop text reading.

JustSayIt preference management wrappers

Index

JustSayIt.set_preferences!

JustSayIt.load_preference

JustSayIt.has_preference

JustSayIt.delete_preferences!

Documentation

JustSayIt.set_preferences! — Function

set_preferences!(args...)

Set preferences for JustSayIt.

This function simply calls Preferences.@set_preferences!, in order to set the package to set preferences for to JustSayIt. For usage documentation beyond the example below refer to the documentation of Preferences.@set_preferences!.

Example

julia> using JustSayIt
julia> JustSayIt.set_preferences!("OPENAI_API_KEY"=>"your-api-key")

JustSayIt.load_preference — Function

load_preference(args...)

Load a preference for JustSayIt.

This function simply calls Preferences.@load_preference, in order to load the package to load preferences for to JustSayIt. For usage documentation beyond the example below refer to the documentation of Preferences.@load_preference.

Example

julia> using JustSayIt
julia> JustSayIt.load_preference("OPENAI_API_KEY")
"your-api-key"

JustSayIt.has_preference — Function

has_preference(args...)

Check if a preference exists for JustSayIt.

This function simply calls Preferences.@has_preference, in order to check the package to check preferences for to JustSayIt. For usage documentation beyond the example below refer to the documentation of Preferences.@has_preference.

Example

julia> using JustSayIt
julia> JustSayIt.has_preference("OPENAI_API_KEY")
true

JustSayIt.delete_preferences! — Function

delete_preferences!(args...)

Delete preferences for JustSayIt.

This function simply calls Preferences.@delete_preferences!, in order to delete the package to delete preferences for to JustSayIt. For usage documentation beyond the example below refer to the documentation of Preferences.@delete_preferences!.

Example

```julia-repl julia> using JustSayIt julia> JustSayIt.deletepreferences!("OPENAIAPIKEY") julia> JustSayIt.haspreference("OPENAIAPIKEY") false